docs-site-project/requirements.txt
Flask==2.3.2
markdown==3.4.3
PyYAML==6.0
Pygments==2.15.1
gunicorn==21.2.0
